Add tests for all latency
